+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## [1.0.1]
 
--   change controller params to private
+- change controller params to private
+
+## [1.0.2]
+
+- change "interpolator" to "curve"